move files around and rename
